Extraído el AlojamientosService
